Moved compile time scanner to core::scan::compile
